requirements.txt
python-dotenv==0.15.0
python-telegram-bot==13.0
requests==2.24.0
